Send a super poke to all connected clients.
We now use two channels per doc.
- One presence channel which is shared between all users on that
  document.
- One private channel which is unique to each user.

The presence channel allows us to iterate over all the users on that
document and we can send a super poke to each of them.

We also store the LastCookie in the Client table so that the client can
determine if it is relevant to them.

The client side has not yet been updated to deal with this.